NaCl: Fix test to accept old, erroneous behavior and new correct behavior pending DEPS roll
The manifest lookup code, when a lookup fails, returns a -1 for the
descriptor and should return a status value which should contain an
errno.  The old code erroneously returned a 0 status, and that got
captured in the golden output unnoticed.  This temporary fix will
accept both the old and new behavior, so that the DEPS roll will not
turn the tree red.

R=sehr@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3318


Review URL: https://chromiumcodereview.appspot.com/12601013