Update tracing framework to optionally use a ringbuffer.
This updates the code to allow passing a continuous_tracing option into the trace framework. When enabled a ringbuffer will be used for event storage instead of a simple vector.

The continuous_tracing option will allow us to trace for longer periods of time without needing to worry that we'll stop tracing before hitting the event we're looking for.

BUG=156025


Review URL: https://chromiumcodereview.appspot.com/12096115