Disable UMA uploads to the XML server.
The XML upload path is deprecated in favor of the protocol buffer path.

For now, this is just the minimal change to disable uploads, so it's trivial to
revert if need be.  I'll follow this up with a larger CL to rip out the old code
once we're sure this change is safe.

BUG=109818
TEST=No more uploads from recent builds to the XML server.


Review URL: https://chromiumcodereview.appspot.com/12716021