Prevented connecting events to a SwappedOut RenderViewHost in WebContentsViewGtk
Added a check to see if the RenderWidgetHost passed to CreateViewForWidget is 
SwappedOut before connecting events to it, or creating a DragDest from it.

Needed for: https://chromiumcodereview.appspot.com/12086095/
When a postMessage is sent to a guest process, like a BrowserPlugin guest,
WebContentsViewGtk::CreateViewForWidget is called with a swapped out 
RenderWidgetHost. This RenderWidgetHost is then used as the WebDragDest for the
window, preventing the window from receiving Drag and Drop events.

BUG= 161112, 177667

Review URL: https://chromiumcodereview.appspot.com/12252016