Fix a crashing call to GURL::spec() in downloads_api.cc:DownloadItemToJSON()
The actual bug is DownloadItemImpl::GetOriginalUrl(): the front() of an empty vector is not a good GURL reference!

The sprinkling of GURL::is_valid() is just to make me feel better.

There is still the open question of what happened to the downloads_url_chain records, but that is less urgent than stopping the crashes.

BUG=190096

Review URL: https://codereview.chromium.org/12607011