Browser Plugin: Implement window.open in guests
Sample API:

function handleNewWindow(event) {
  // The default behavior is to discard the new window. If we don't prevent
  // default or event.window is no longer reachable the garbage collector will
  // automagically discard the window for us.
  event.preventDefault();
  chrome.app.window.create('newwindow.html', {
    top: 0,
    left: 0,
    width: 640,
    height: 480,
  }, function(newwindow) {
    newwindow.contentWindow.onload = function(e) { 
      var newwebview = newwindow.contentWindow.document.querySelector("webview");
      event.window.attach(newwebview);
    }   
  }); 
}

webview.addEventListener('newwindow', handleNewWindow);

This is implemented by reusing a lot of the existing code for the permission API.
The event.window object is managed by the v8 garbage collector.

The new BrowserPluginGuest's lifetime is managed by its opener as long as it's
not attached. Once a BrowserPluginGuest is attached, it has an embedder
WebContents, and loading of resources begins.

BUG=140316
Test=WebViewTest.Shim { webViewNewWindow, webViewNewWindowTwListeners, webViewNewWindowNoPreventDefault, webViewNoReferrerLink }

Review URL: https://chromiumcodereview.appspot.com/11280291