Allow for synchronous compositor input handling when UI and compositor threads are merged.
In particular, we provide a hook for input event filtering in 
RenderWidgetHostView, which is used by RenderWidgetHostImpl during input event 
processing. When the UI and compositor threads are merged (via a 
WebView-specific flag), this hook allows compositor-relevant input events to be 
filtered and processed synchronously, prior to their being forwarded to the 
renderer. 

BUG=

Review URL: https://chromiumcodereview.appspot.com/12416005