Add PPAPI C++ wrappers for PDF.
Add C++ wrappers for PDF. The wrappers aren't entirely idiomatic (e.g. there
is a case where we just return a PP_Resource instead of wrapping it in an
object). This is because of the weird way that the interface is used in the
plugin which just makes it more convenient to return the raw resource ID.

BUG=


Review URL: https://chromiumcodereview.appspot.com/12527012