Update TestURLRequest constructor interface in unit tests.
The webkit and android_webview unit tests that create TestURLRequest objects
are updated in this patch to explicitly specify which network delegate they
use, rather than extracting it from the context.  This is usually but not
always NULL.

This change is motivated by a desire to remove external accesses to a request
context's delegate. (If a block of code needs to use a NetworkDelegate, it
should not query the URLRequestContext for it, but instead should already have
access to whatever delegate the context uses.)

BUG=146587


Review URL: https://chromiumcodereview.appspot.com/12321106