Removed support for Core Animation plugins when accelerated compositing is disabled. CA plugins have been rendered through Chrome's compositor for several releases now and it has not been found necessary to revert back to the older behavior.
If the compositor is disabled, then support for the Core Animation and Invalidating Core Animation drawing models is no longer advertised.

Tested with the Unity and NPAPI Flash plugins. As expected, Unity no longer renders with the compositor disabled; NPAPI Flash falls back to the Core Graphics drawing model. (Pepper Flash is not affected; it happens to require the compositor before and after this change.)

BUG=105344


Review URL: https://chromiumcodereview.appspot.com/12641008