Android: fix DEPS to work with git submodules.
Android depends on some of the unix DEPS entries as well as the android
DEPS entries. gclient checks out both, but git crup/crsync don't. Add
the dependencies that the android port needs, using a variable to make
sure the same revision is used for linux/android. gclient still does the
right thing with this change as long as the two DEPS entries point to
the same repository (which they do).

BUG=

Review URL: https://codereview.chromium.org/12716020