Add policies for periodic reboot and reboot-after-update
This CL adds two new device policies: Reboot after X amount of uptime and
reboot when a reboot is necessary to finish applying an update. Such
automatic reboots may be delayed by up to 24 hours while Chrome waits for
a moment that causes the least disruption to the user (nobody logged in,
nobody interacting with the login screen). If no convenient moment comes
for 24 hours, a reboot is forced.

The implementation persists information so that reboots happen at the
expected times even if Chrome crashes in between. There are safeguards
against reboot loops due to corrupted information.

BUG=chromium-os:36400,177029
TEST=Extensive unit test coverage


Review URL: https://chromiumcodereview.appspot.com/12578005