Add support for WEBKIT_USE_NEW_WEBFILESYSTEMTYPE
Upstream WebKit is migrating from WebFileSystem::Type to a new top-level WebFileSystemType (unconnected to the WebFileSystem class) in order to make other filesystem-related refactoring easier. This is behind an #ifdef so that we can make the necessary changes upstream and downstream all at once.

Review URL: https://chromiumcodereview.appspot.com/12886018