cc: Add --ui-foo flags for all compositor debug visualizations.
For every --show-debug-thing flag, we should have a mirrored
--ui-show-debug-thing flag to show the same information in the
UI compositor.

Move the current ui-compositor flags to cc/base/switches.h where the
renderer-compositor flags are, so that the mirrored nature is easily
seen and maintained. Also move the few renderer-compositor flags for
these compositor features to cc/base/switches.h.

The only flag that can't move is kShowPaintRects, as it is also used
for the software non-compositing path. So instead we mirror this flag
in the ui-compositor switches.

R=piman

Review URL: https://codereview.chromium.org/12909003