Port PerfWarm startup test to Telemetry
First step of porting tests in chrome/test/perf/startup_test.cc to Telemetry.  The motivation for this is to lower flakiness and make it easier to create startup tests using dirty profiles.

Note: The test measures the time it takes the browser to start spinning the main message loop so the exact URL used for the test is irrelevant.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/12770005