[Sync] Plumb local favicon deletions to sync favicons datatype
If the user deletes all the local visits to a url, triggering a favicon
expiration, we now delete the favicon in sync too. Similarly, if the user
clears all their browsing history, we delete all synced favicons.

This involves tracking which favicon icon urls are expired during history
deletes and including that information in the URLsDeletedDetails.

BUG=154886

Review URL: https://codereview.chromium.org/12779025