Pepper: Use shared memory for ArrayBufferVar.
This change improves performance for sending ArrayBufferVars between the
plugin and the host. It copies the data from the var into a shared
memory region and sends the shared memory handle over IPC.

This required defining a new message for plugins to request the host to
create shared memory.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11827059