[Android] Allow JNI initialization on background thread
The code used to insist that JNI initialization only happened on
the main (UI) thread. This was because it was believed that
the static initialization of the Java classes initialized as
a result of JNI calls might only work on the UI thread. In
practice it seems that none of these Java classes have this
requirement, so this seems to have been an unnecessary
restriction, and slowed up startup on Chrome for Android. 

Remove this restriction, but add locks to ensure that only one thread is trying to initialize the JNI at a time.

Also change some function names to reflect this change.

BUG=214560


Review URL: https://chromiumcodereview.appspot.com/12700011