Parsing filesystem url before giving it to media player
Android media player directly takes a url to playback a video and it cannot handle filesystem urls.
To solve the problem, we need to translate a filesystem url to a platform path before calling setDataSource().
CookieGetter is renamed to MediaResourceGetter to handle such translations.

BUG=180541


Review URL: https://chromiumcodereview.appspot.com/12595005