Disabled notification collapsing.
Temporarily disabled the creation of collapsed notifications that have
an expand button because those might be causing layout issues when
their size is changed by a transition from collapsed to expand.

All notifications in the system (in the notification center as well as
in toasts) are now always expanded.

BUG=NONE
R=mukai@chromium.org


Review URL: https://chromiumcodereview.appspot.com/12942003