PPAPI: Remove ImageDataInstanceCache on instance shutdown
Remove the ImageDataInstanceCache when the instance gets destroyed. This fixes a DCHECK (see the bug).

(Note that even before this change, The ImageData resources themselves time out after 2 seconds, so aren't leaked. The ImageDataInstanceCache is deleted when it has no more ImageDatas, so it also is not leaked.)

BUG=222159


Review URL: https://chromiumcodereview.appspot.com/12440053