Limit clustering to 10 until run_test_cases gets better.
We are working on having run_test_cases parse the output on the fly. Until this
is done, limit the clusters to 10 ten cases. This value is taken from empirical
worst case of browser_tests behavior, so that a single cluster doesn't take more
than 6 minutes.

R=csharp@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/12962005