Fixed overflow bug that was causing layout problems with notifications
It turns out views doesn't like overflows so it's not a good idea to
return std::numeric_limits<int>::max() for the preferred width of a
view that wants to occupy all of its parent's width :-).

Now returning message_center::kNotificationWidth, which is fine because
the parent of this particular view will never be wider than the fixed
notification width.

BUG=222178,191143,222108

Review URL: https://chromiumcodereview.appspot.com/12491020