Correcting the anchor offset calculation
This is a revert of CL 12498004 since it produced a tiny - not really noticable offset shift of the first icon.
With the possibility to move icons around it became apparent since the first icon can now be a rectangular item.
To address this, the old spacing trick had to be applied again and the anchor point for menus and tooltips had to be adjusted accordingly.

That said - as soon as Simon's change lands, the first item needs proper handling again since the first item can change. 

BUG=218373
TES=visual


Review URL: https://chromiumcodereview.appspot.com/12660015