cc: Chromify LayerTreeTest.
This renamed ThreadedTest -> LayerTreeTest, and
layer_tree_test_common.h to layer_tree_test.h so that they
match. I moved the other classes from the header to the
layer_tree_test.cc file.

The names of the classes are cleaned up. MockFoo classes that
are not mocks are renamed to FooForTesting.

The member vars of the LayerTreeTest class are made private
with accessors. 3 tests in LayerTreeHostTests depended on this
access, but were not testing anything, so I've removed them.

R=enne
BUG=

Review URL: https://codereview.chromium.org/12648010