Make power API extension set a default icon and title.
This makes the chrome.power "Keep Awake" example extension
define a default icon and title in its manifest file.
Otherwise, the extension works when it's first installed but
appears to die with "This extension has no action specified"
errors on subsequent runs.

(It also makes the extension add a click listener before
attempting to update the title and icon.  Perhaps I was just
encountering a race, but making this change also seemed to
work around the problem.)

BUG=none

Review URL: https://codereview.chromium.org/12827011