Added survey butter-bar to web-app.
The butter-bar is shown until the user either closes it with the X icon or begins the survey. Once dismissed, a cookie is stored in sync storage to prevent it showing up again for that user on another computer (note that this uses Chrome Sync, so it relies on them being signed in to Chrome).

BUG=191368

Review URL: https://codereview.chromium.org/12566035