Use timeout that is reset everytime a test case completes.
This removes the timeout based on the number of test case per cluster. This was
problematic because for large clusters, the timeout was becoming so large that
it was useless.

Now the timeout is kept to the same value and is reset everytime a test case
completes.

This will permit using much larger test case clustering.

R=csharp@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/13010002