Suppress touchscreen tap immediately after a GestureFlingCancel
Tapping on the touchscreen during a fling scroll causes a
GestureFlingCancel to be sent to stop the fling. It also sends a pair of
GestureTapDown/GestureTap events that should be suppressed if the fling
is successfully stopped (otherwise we will have undesirable side
effects; e.g. activating a link or push button in tap location). This
was working for touchpad. It is now added for touchscreen, too.

BUG=162242


Review URL: https://chromiumcodereview.appspot.com/12087140