Added "None available" title and grey out the media menu when there is no device.
When capture device list is empty, we show a "None available" title to clearly tell the users there is no device, instead of having empty menu.
This will also fix the problem when the height of the empty menu is not the same as the other menu with item.

BUG=178383

Review URL: https://codereview.chromium.org/12457016