Remove dead method WebContents::ShouldAcceptDragAndDrop
This method is no longer used. It was originally intended to prevent
drag and drop operations on popup windows and panels on CrOS from
triggering navigation operations in a new tab. Unfortunately, this is
not the correct way to fix the problem anyway; if this method were
actually used, it would prevent *all* drag and drop from working in the
popup/panel window. The correct fix is to plumb through the relevant
drag and drop policy to WebKit so the renderer knows not to initiate a
navigation.

BUG=none
TBR=jam

Review URL: https://codereview.chromium.org/12938013