Send Feedback Experiment
Feedback Experiment to test response rate (usage of report an issue) when alternate text (Send Feedback) and or an alternate location (on windows/linux/mac the main Wrench Menu, on ChromeOS in the more tools option) is used. In the case of alternate text experiments the label for the report an issue button on the chrome://help page is also changed.


BUG=169339
TEST=Manual
Default behavior should not change, to confirm desired behavior run chrome with --force-fieldtrials=SendFeedbackLinkLocation/$GROUP_NAME/
Where GROUP_NAME is one of {alt-text,alt-location,alt-text-and-location} (note trailing '/' is required!)

Review URL: https://chromiumcodereview.appspot.com/12490012