Windows: Fix bug in interpretation of return code for win32 registry functions.
RegOpenKeyEx and friends return non-zero on error. The FAILED macro is not the right way to detect failure.

The DisplayLinkVersion function actually still works with this bug because even if HKLM/SOFTWARE/DisplayLink/Core does not exist, reading the value of HKLM/SOFTWARE/DisplayLink/Core/Version results in empty string which results in an invalid version. This is what DisplayLinkVersion should return if DisplayLink is not installed.

Review URL: https://chromiumcodereview.appspot.com/12913007