Replace edit_url by resource_id for AuthorizeApp.
This is preparation of the temporary implementation of AuthorizeApp for
Drive API v2. Unfortunately authorizing third party application is not
supported on Drive API v2 for now, but we probably can use the GData WAPI's
one as a very short term work around. Also, it should be supported somehow
soon.

BUG=179837
TEST=Ran unit_tests and tested on chromeos-chrome manually.


Review URL: https://chromiumcodereview.appspot.com/12989014