Reland libjingle.gyp changes. The changes are the same as in the previous review: https://codereview.chromium.org/12457027/
I had to revert due to build problems but eventually tracked
them down to chromeos dependencies in sync.  For 'sync', this
is a problem for shared_library builds but not otherwise.
For sync_unit_tests, this is a problem since sync_unit_tests
is always built as an executable.

Previously these problems were masked since the dependencies
were being pulled in via libjingle.

TBR=ronghuawu,akalin

Review URL: https://codereview.chromium.org/12930010