Expose a concept of bottom overdraw amount for cc drawing viewport.
When the physical output surface is larger than the visible viewport,
we need to subtract the overdraw to ensure you can scroll to the
full content of the page.

Resizing the output surface is expensive on mobile devices, so we
minimize resizes and deal with this size mismatch.

BUG=161303, 221995

Review URL: https://chromiumcodereview.appspot.com/12437023