Limit to only 10 image requests per page in ResourceScheduler.
The pending requests are now kept in a net::PriorityQueue. Priority changes
affect the order of the queue. After painting or network idle, we'll load up to
10 images simultaneously in PriorityQueue order.

This is a 2% improvement in Speed Index, 4% in first paint, and 5% in
DOMContentLoaded. There is a 0.5% regression in onload.


BUG=None


Review URL: https://chromiumcodereview.appspot.com/12874003