Add hooks to allow retrieving the current known tracing categories.
We want to be able to select the categories to be traced before
we generate the trace file. This CL exposes the needed hooks
for about:tracing to query for the currently registered tracing
categories.

The TracingMessageHandler will listen for 'getKnownCategories'
from about:tracing and, when complete, will call
'tracingController.onKnownCategoriesCollected' and pass the
current categories back to about:tracing. If we are unable
to perform the request to get categories we will call
'tracingController.onKnownCategoriesCollected' with an empty
list of categories.

BUG=191463


Review URL: https://chromiumcodereview.appspot.com/12859004