TextureImageTransportSurface: Reuse TextureDefinition
This passes around ownership of the TextureDefinition between
mailbox manager and surface. It will make it easier to switch to
a refcounted scheme for mailbox textures.

I optimized Texture::Save()/Restore() just because when I first
create the TextureDefinition in ImageTransportSurface, I do not
have a TextureManager yet to figure out what the max. number of
needed levels is.

So I added a bunch of unit tests for that.

BUG=181640,220561


Review URL: https://chromiumcodereview.appspot.com/12893006