[win] Fix flickery title bars in packaged apps.
This patch works around a "feature" of WM_NCCALCSIZE where if you specify that
the client rect should be equal to the window rect, it changes into what is
presumably some sort of backwards compatability mode and decides that it will
be flickering your window content between opaque and transparent now.
Sometimes.

This patch tells Windows the client rect is actually 1px taller than the
window, then lies to user code above it about what the client rect is so that
the bottom 1px of the content doesn't get cut off.

BUG=132948
R=msw@chromium.org,sky@chromium.org,ben@chromium.org


Review URL: https://chromiumcodereview.appspot.com/12981003