Show an InfoBar when trying to start Packaged Apps from Metro mode.
Platform Apps do not function properly while in Metro mode.

This change intercepts platform app launch requests done
while the browser is in metro mode, redirecting them to an
infobar. If "yes" is chosen, a local pref is stored with the
extension id and Profile that tried to launch it, and a relaunch
of Chrome in Desktop mode is triggered. Upon restart,
apps::AppLaunchOnRestartService is responsible for launching the
selected app, and clearing the pref.

BUG=153426
TEST=With Chrome configured for Windows 8 mode, try to start a
packaged app. This can be done from NTP (if launcher not enabled),
or from a taskbar shortcut, a desktop shortcut, or a pinned start
page tile. InfoBar should display offering to switch to Desktop mode.
If switching, selected packaged app should launch after a brief
delay.

Review URL: https://chromiumcodereview.appspot.com/12450014