OSX App List Pager and root view controller.
Adds a root AppListView for the OSX app list, and moves ownership of 
the AppListViewDelegate there (as well as principal ownership of the 
GridController). The root view contains the apps grid, a pager control, 
and a background. An AppsPaginationModelObserver feeds scroll changes 
on the grid to the view controller to update the pager (an 
NSSegmentedControl). Clicking the pager scrolls to the corresponding
page in the apps grid view.

BUG=138633
TEST=Added app_list_unittests: AppListViewControllerTest.*, AppsGridControllerTest.PaginationObserver*


Review URL: https://chromiumcodereview.appspot.com/12912004