In the tab capture code, use the FrameSubscriber mechanism and the notification service to learn of paint activity, and to initiate frame captures only when the underlying content changes.
The control logic concerning when and how to initiate captures is moved from the "old" CaptureMachine into a new class called the CaptureOracle, which is the only multi-threaded object in the system. The CaptureMachine lives on the UI thread, and is now responsible mostly for tracking the active RenderWidgetHost and starting manual copies from it (the BackingStoreCopier is subsumed by the "new" CaptureMachine). For the active RWH, the CaptureMachine installs a ContentCaptureSubscription which routes events (meaning paints and timer events) from their origin to the Oracle.

The old SynchronizedConsumer class is subsumed by the CaptureOracle. The Oracle has something called a SmoothEventSampler which handles the case where content is being produced faster than our target frame rate. Its implementation is dumb, but seems to work in the common cases.

Parts of the old CaptureMachine become the WebContentsVideoCaptureDevice::Impl. I think the separating is pretty self-evident. This class no longer needs to be deleted from an outside thread, as it doesn't run its own "manager thread" anymore. A couple subtle problems are fixed in the way errors are handled by this class.

BUG=181814,223427,181594
TEST=content_unittests, manual testing with extensions

Review URL: https://chromiumcodereview.appspot.com/12731011