Add elevation to the managed mode navigation observer.
This changelist adds an elevation flag to the managed mode navigation observer
which can be used to save elevation for a certain WebContents. When navigation to a page which is not the Webstore, Extensions, Settings or History page, the elevation will be reset automatically.

BUG=222364
TEST=Manual


Review URL: https://chromiumcodereview.appspot.com/12413028