Use GData WAPI's third party app authorization operation for Drive API v2.
Unfortunately, it is not supported to authorize third party app on Drive API v2.
So, as its workaround, we use GData WAPI's authorization operation.
This should work on Drive API v2, because both protocol should share the
same resource id for most cases.
Note that enabling Drive API v2 with a flag is still under-impelementation, so
now it not ready to test it manually. It will be done when the Drive API v2
starts to work. Note that the code is not yet used, so adding the code
with unit_tests should work (i.e. unless it breaks compilation nor test).

BUG=179837
TEST=Ran unit_tests


Review URL: https://chromiumcodereview.appspot.com/12924009