Fix accessible bounding box when page is scrolled.
This patch is for merging; will add new test to prevent
this from happening again in subsequent changelist.

See this diff from the changelist that broke it to see the bug:
https://chromiumcodereview.appspot.com/12084028/diff/12004/content/browser/accessibility/browser_accessibility.cc
Originally the scrolling logic was applied to the main frame. Change 12084028 intended to have it affect all frames, but instead it affected all frames except the root frame (which has a distinct role in the accessibility tree).

BUG=223756
TBR=dtseng

Review URL: https://codereview.chromium.org/13007004