ibus unit tests: Do bool checks using EXPECT_TRUE instead of ASSERT_EQ
This is a workaround until we can upgrade to a newer version of
gtest on Chrome OS.

The version of gtest used on Chrome OS doesn't correctly
handle EXPECT_EQ with boolean arguments. In particularly the unit tests
fail to compile with the new tool chain due to [-Werror=conversion-null]
warnings.

BUG=none
TEST=built chromeos_unittests using new tool chain


Review URL: https://chromiumcodereview.appspot.com/12634032