Bluetooth UI: Don't close the overlay on "connect" action.
The connect button in the pairing overlay
('bluetooth-pair-device-connect-button') closes the overlay without changing
the pairing status to any of the final states (Dismissed or ConnectFailed).
This causes the pairing to be canceled sending a "CancelPairing" call to the
BlueZ backend.

This fix keeps the pairing dialog open until a notification from the chromium
option_handler is received.

BUG=chromium:223308
TEST=Attempt to pair a keyboard with a pin code, fail once and try again. The second time the chromebook will ask for a pincode and it should work when entering both pincodes.


Review URL: https://chromiumcodereview.appspot.com/12703027