Replace only use of WebRegularExpression with RE2
WebKit::WebRegularExpression is a wrapper for WebCore::RegularExpression which
is in turn a wrapper for YARR, a regular expression engine in JavaScriptCore.
We added a dependency on WebKit::WebRegularExpression in r71498 to match
style attributes that match the width/height of blocked plugins so when plugins
are removed from a page transparent placeholders that eat events aren't left
behind.

As it turns out, this is the only user of WebCore::RegularExpression in all of
chromium that is reachable at runtime (there's some code in WebCore::Frame that
is compiled in but not called by us that uses WebCore::RegularExpression).

R=darin@chromium.org


Review URL: https://chromiumcodereview.appspot.com/12741011