remove redundant DCHECK that a size_t variable >= 0
SampleVector::GetCountAtIndex has a DCHECK which checks that a size_t
variable is >= 0 and then some other condition.  But size_t is an
unsigned type so the first part of the check is useless (and it
gives an error because of this with clang).


Review URL: https://chromiumcodereview.appspot.com/13006014