Alternate NTP Mac: Push down overlay in presentation mode
In presentation mode the browser toolbar and bookmark bar are in a
special floating view. This view obscures the instant overlay since
the overlay is at the top of the screen.

This CL pushes the overlay down so that it's not obscured when the
floating toolbar view is shown.

Test build: https://docs.google.com/file/d/0B0Odde3V7EhWTmtMTXBJYmtGS3M/edit?usp=sharing

BUG=167185
TEST=
  Launched Chrome with instant extended enabled.
  Enter presentation mode and type in the omnibox.
  Verify that the omnibox popup is below the floating toolbar view.

Review URL: https://chromiumcodereview.appspot.com/12594015