Adjust buildCache per builder to be a multiple of builder's slave count.
This patch introduces an auto-adjustment parameter specifically designed for the tryserver where each builder's cache is adjusted to a multiple of its slave count.

Buildbot's buildCacheSize is a single number for all builders. The tryserver's main caching requirement is that builds CQ is polling are cached in memory. A fixed buildCacheSize means that builders with a low number of slaves may consume as much cache as those with a large number of slaves (and thus, CQ builds). Adjusting the number too low means Buildbot spends its time unpickling from disk, and setting it too high means it may grow past the memory limits of the system. This patch addresses this problem.

Context and background for this change is discussed in https://chromiumcodereview.appspot.com/12990005/

BUG=222492

Review URL: https://chromiumcodereview.appspot.com/12990005