android: Update comment about invalid timestamps from SurfaceFlinger
A frame with a timestamp of INT64_MAX actually means that the fence
object associated with the frame is still pending or that the frame has
not yet been displayed. Update the comment in SurfaceStatsCollector to
make this more obvious.

No functional changes.

NOTRY=true


Review URL: https://chromiumcodereview.appspot.com/13188002