Make sure the delegate view always informs the renderer when a drag ends.
There are several instances where a renderer can attempt to start a drag
but it fails on the browser side. If this happens, the browser must
inform the renderer that the drag has ended; otherwise, the renderer will
end up in a state where it ignores all input.

BUG=157134
TBR=joi

Review URL: https://codereview.chromium.org/13008031