Fix Android-specific shmem size code.
In the past, this OS_ANDROID block assigned a value to created_size_
(which appears not to have been used for anything), whereas mapped_size_
used the cross-platform codepath.  In r191098 jschuh@ disabled the
normal assignment of mapped_size_ on Android and made this old code set
the value of mapped_size_.

This Android code block was there to set "bytes" in when Map is called with
a size argument of 0.  So after r191098, mapped_size_ was never set when the
argument was nonzero.  Bring back the mapped_size_ setting and
clean up the Android block.

NOTRY=true
BUG=224646


Review URL: https://chromiumcodereview.appspot.com/13133004