Set pepper flash fullscreen window to opaque.
CGLFlushDrawable was taking a really long time when presenting to a fullscreen pepper flash window when using a discrete GPU while the Chrome window was also being drawn to. Setting the window to opaque seems to fix that, and still works because the GL content is always drawn on top of the Flash window, and not underneath.

BUG=171911


Review URL: https://chromiumcodereview.appspot.com/13219004