Fix accessibility crash due to reparenting an iframe.
The bug occurred because RendererAccessibilityComplete assumed incorrectly
that WebKit will never reparent a WebAccessibilityObject. This is true within
one document, but not true when iframes are involved - when an iframe moves,
the iframe's accessible document objected is reparented to the new iframe
object. The crash happened because sanity-checking code in
BrowserAccessibilityManager detected the error and triggered the crash.

The fix handles reparenting in the general case. RendererAccessibilityComplete
keeps track of the parent of each node in its tree, and when it sees the
parent is about to change, it just sends an update to remove the child from
the old parent strictly before the update to add the child to the new parent.

Repro: clear your cache and visit http://en.espnf1.com/ with the
--force-renderer-accessibility flag. Without this patch it will crash.

BUG=222212


Review URL: https://chromiumcodereview.appspot.com/13207002