[NaCl SDK] nacl_io: Fix bug when fseeking to end of httpfs-mounted file.
If the server does not set a "Content-Length" header, the MountHttpNode did not
know the content size until the first read occurred. This means that an fseek()
before the fread() will fail; the "end" of the file is 0.

This CL makes MountHttpNode::GetSize() perform a Read() if the content size has
not yet been cached.

BUG=none
R=sbc@chromium.org
NOTRY=true


Review URL: https://chromiumcodereview.appspot.com/13327007