Fixed GetContentsBounds() so GetInsets() can be overridden.
View defines GetInsets() as virtual but that method could not be safely
overridden because the implementation of the non-virtual
GetContentsBounds() duplicated some of the logic of GetInsets().

This changes GetContentsBounds()'s implementatation to call GetInsets()
so overridding GetInsets() changes GetContentBounds()'s behavior in line
with changes to GetInsets().

BUG=none
R=sadrul@chromium.org


Review URL: https://chromiumcodereview.appspot.com/12887012