Capture a hidden RWHV on Mac when NSWindow is not available
When a RWHV is being captured it is forced shown, however there
is no NSWindow for drawing so it drops the call to DrawIOSurface.
Frame subscriber however needs a DrawIOSurface to capture the
current frame, the consequence is that tab capturing fails.

This change manually does a copy for the frame subscriber if
NSWindow is not available.

BUG=223988


Review URL: https://chromiumcodereview.appspot.com/13140002