Pass spellcheck request identifier to the spelling service client
Pass spellcheck request identifier to the spelling service client instead of
storing it in the spellcheck message filter. This avoids some issues with using
outdated spellcheck results. Note that the spelling service client does not 
access the request identifier. The identifier is stored as part of the 
spellcheck-finished callback.

BUG=179639

Review URL: https://chromiumcodereview.appspot.com/13330004