cc: Use highp precision for texture coords if available and needed
This fixes an earlier version of this patch that had a bug in
FragmentShaderRGBATexAlphaMask.

High precision is only used when the source is greater than N
pixels in any direction to prevent performance regressions. N is
determined based on the actual precision of mediump reported by
the driver.

BUG=183581
BUG=173747

Review URL: https://chromiumcodereview.appspot.com/12665005