Convert native strip from rule to action
First, this allows us to correctly have the output of the copy/strip
action be an input to the apk package action.

Second, this step now uses the list of libraries created by ordered
libraries step. This is required to support the component build where
we can not manually list the required libraries.

BUG=158821


Review URL: https://chromiumcodereview.appspot.com/13058003