Extract metadata for media elements
In our current html5 media implemtation, we pretend everything is loaded and gave a temporary duration of 100 seconds to WebKit.
However, this will break things if developer do sth like video.seek(video.duration). Also, width and height information are not available.
This change adds a thread in browser process to extract media metadata from the elements, before reporting to WebKit that media is loaded.
This only works if wifi or ethernet is available.


Review URL: https://chromiumcodereview.appspot.com/12518035