Initialize the simple cache backend at runtime.
We will need to choose the backend at runtime for A/B testing. This change is 
the first step towards backend comparisons: it moves the logic of choosing among 
cache backends to a single point (cache_creator.cc) and makes it logically 
separate from BackendImpl. Next step would be to inject performance-analyzer 
backend into this cache backend choice logic. The analyzer is going to be under 
a compile-time flag. 

With this change the Simple Cache Backend operation gets enabled when the 
"SimpleCacheTrial" experiment is set.

Cache Initialization. 

The Simple Cache Backend now shares the common initialization/cleanup code. So 
it can delete inconsistent cache, cache of a different version etc 
asynchronously. It is just a positive side-effect of moving the backend choice 
to a single point. 

The CacheCreator with almost no change. 

To see the difference:
shell> out/Debug/chrome --enable-logging=stderr --v=1 \ 
  --user-data-dir=/tmp/unique --use-simple-cache-backend=on \ 
  http://url 2>&1 | grep "Simple Cache"

BUG=173390, 173384
TBR=maruel

Review URL: https://chromiumcodereview.appspot.com/12794003