[Spellcheck] Add separator in front of suggestions
If there are any suggestions available, add a separator in front of
it so they are visually distinct from earlier entries in the menu.
(I.e. link-related entries)

This is safe to do even if the suggestion is the first entry, since
AddSeparator only adds separators if they wouldn't be the first menu
entry.

TBR=brettw@chromium.org
R=rouslan@chromium.org
BUG=224687


Review URL: https://chromiumcodereview.appspot.com/13258004