Do not show "Ask google for suggestions" in panel's context menu
Selecting "Ask Google for suggestions" has caused some kind of bad side-effects. On Windows, an info dialog that is limited to the panel shows up and it looks bad on small panel-window (this was fixed on the main trunk but the change is not in M27). On Linux, it tries to use the bubble and causes a crash. We will put it back once the crash issue is resolved.

BUG=224807
TEST=Manual test by opening a panel and bringing up context menu on a textbox


Review URL: https://chromiumcodereview.appspot.com/13416002