Make GL_TEXTURE_EXTERNAL_OES texture immutable.
This is semi-redundant. The reason things work now is there are
2 GLenum validators. One for glBindTexture and a different one
for glTexImage2D etc. The one for glTexImage2D does not have
GL_TEXTURE_EXTERNAL_OES in it.

But, It seemed best to mark the texture as immutable just in case
someone forgets. Also, it seemed best to add a couple more tests
that test that things work.

In that vain it was clear a few things that were public (npot,
texture_complete, cube_complete) should not have been public
so fixed that.

BUG=none


Review URL: https://chromiumcodereview.appspot.com/13463010