Split the power monitoring feature from SystemMonitor
The SystemMonitor is a mixed monitor which not only monitors
the power state changes but also the devices changes. This
patch is to separate the power monitor from SystemMonitor
as a new class PowerMonitor which is dedicated to monitor
power state.

The next step is to seek a opportunity to refactor SystemMonitor
as something like DeviceMonitor.

BUG=149059
TEST=base_unittests --gtest_filter=PowerMonitorTest.*

Review URL: https://chromiumcodereview.appspot.com/10959020