CC: HUD layer shouldn't cause damage when --show-surface-damage-rects
Having the HUD layer cause damage defeats the purpose of showing the damage rects
(it shows the full frame as damaged). However, we want all of the HUD layer to
display. So we drop the HUD layer damage and disable partial swaps when a HUD
layer is active.

BUG=177724


Review URL: https://chromiumcodereview.appspot.com/12924010