Add dropdowns to autofill dialog
Adding the dropdown for country, month and year with the glue that
gives the related string array for a given field. Made changes so
the dialog calls are aware that some views in the edit layout are
spinners and not edittexts. This gives a non-null UI element for each field
now.

BUG=181538
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/13446004