Fix missing scrollbar thumb on WinXP.
It turns out that the XP theme, the default desktop theme
on WindowsXP, draws the scrollbar thumb and its grippers with an
alpha value of zero and therefore they become invisible. This patch
sends them down the "indirect draw" path that paints them to an
off-screen canvas and fixes their alpha values.

I considered sending all scrollbar draws through the indirect path
since there could in theory be other themes that suffer from the
same issue but decided against it given the inefficiency of the
indirect drawing path. At this point the only theme that seems
to have this issue is the XP theme which is only available
on WinXP. Note that the Windows Classic theme which is available
on later versions of the OS is different from the XP theme and is
already going down the indirect draw path. 

The Windows Classic theme is the only theme for which 
GetThemeHandle(SCROLLBAR) returns NULL.

BUG=179643


Review URL: https://chromiumcodereview.appspot.com/13649002