Omnibox: Don't Allow Navsuggestion (even from cached results) in Keyword Mode
This is an improved version of
https://chromiumcodereview.appspot.com/13031006

Here is the changelist description for that commit.  The description
still applies.
---
In keyword mode, demote all navsuggestion results below the best ranking
keyword query mode.  This is necessary so a navsuggestion doesn't appear
inline, which would end up kicking the user out of keyword mode.

This has to be done in SearchProvider.  It cannot be done on the suggest
server side because the suggest server does not know whether it is
serving results to a default search engine or a keyword search engine.

Note: there can be bad cases when the default search engine and the keyword
search engine both supply relevance scores.  I haven't thought those these 
issues yet as we only know of one engine that supplies scores at this
time.  (And if the same engine is used for the default and keyword search
engine then it's only queried once, so these concerns don't apply.)
---

This new, improved version enforces the constraint latter in the
processing that the current version.  The issue with the current
version is that a navsuggest result could end up appearing first
if it's a result from a previous suggest query and passes the is-not-stale
test (and the query result that previously appeared higher does not).
This is because the constraint used to be enforced during
ParseSuggestResults() which only gets called, uh, after suggest results
have been fetched, being the constraint-fixing logic did not run
for results served before suggest results are returned.

This new placement of the enforcing logic applies all the time and
is more consistent with the location of other enforcing logic.

I should've listened to Mike all along...

By the way, the last paragraph in the original changelist description still
applies.

BUG=196550


Review URL: https://chromiumcodereview.appspot.com/13393013