Screenshot effect non-obvious
Added new (rich) notifications when a screenshot is taken on
Chrome OS. There are two different notifications, one for
success and one for failure. In case of success, the user
can click on the notification to bring up the file
manager with the screenshot file hi-lighted.
I've attached screenshots to the BUG.
Added unit test for the screenshot taker.

BUG=150017
TEST=manual; for now run with --enable-rich-notifications.

Review URL: https://chromiumcodereview.appspot.com/13105002