Enable CompositingIOSurfaceTransformerTest.*
Due to weirdness with GPUs on trybots causing flakiness in tests, the initial implementation of CompositingIOSurfaceTransformer was committed with all unit tests disabled.  This change will enable the tests, but force the software-based OpenGL implementation to be used.

In a later change, we will allow the OpenGL implementation to auto-detect and utilize a GPU, where one is available.  However, first the tests should prove themselves as stable when not using GPU.

TBR=kbr


Review URL: https://chromiumcodereview.appspot.com/13573004