History Action Pop-ups incorrect behavior when document is scrolled.
The Menu Button ui control that the history action pop-ups employ
handle the resize event by hiding the menu.  This CL will similarly
hide the menu on a |scroll| event is triggered.

This resolves the scrolling problem in the referenced bug.

However, this change does not address the reported behavior that the
pop-up remains visible when changing tabs. I tend to agree with
sergiu@ that the existing behavior seems proper.

BUG=222313
TEST=The MenuButton ui element is used in several places; I tested the History, Bookmark Manager and NTP but I can see it also being used in ChromeOS login screen and File Manager which I am unable to test. The change involves hiding the menu pop-up when the document scroll event is triggered.

Review URL: https://chromiumcodereview.appspot.com/13532014