drive: Make FileManagerBrowserTest more robust
Previously, getFileList() was called directly from testFileDisplay().
This was bad, as it's not guaranteed that the file list is already rendered
by the time testFileDisplay() is called, hence the test was flaky.
We should use waitForFileListChange() instead.

Likewise, doKeyboardTest() and doKeyboardTestWithConfirmation() are fixed
in the same way.

BUG=224534

Review URL: https://chromiumcodereview.appspot.com/13450004