Show only the settings pages for new managed users
Currently when creating a new managed user we open other tabs in addition to
the managed users settings one, which we don't want to do.

This makes sure that only that tab gets added to the list of URLs to open on
start-up by early-returning from that function.

R=ben@chromium.org
BUG=171370


Review URL: https://chromiumcodereview.appspot.com/13468005