[mac] Make app shims look in the correct user data dir on Canary.
On Canary, the user data directory is affected by
[[NSBundle mainBundle] objectForInfoDictionaryKey:@"CrProductDirName"].
Unfortunately in app shims, [NSBundle mainBundle] refers to the app shim's
bundle, not Chrome's bundle. This patch parameterises the
user-data-directory-figuring-out code on the NSBundle, so that app shims can
pass an NSBundle for the real Chrome bundle.

R=thakis@chromium.org


Review URL: https://chromiumcodereview.appspot.com/12663023