Host-side third party token validation
This creates a TokenValidator implementation on the host, that upon receiving a token:
Signs the token with its private key.
Uses URLFetcher to request the exchange of the token for a secret from the Token Validation URL.
On receiving a reply, checks that the scope in the reply matches the one required for this connection.
Uses the callback to send the shared_token back to the authentication layer.

(The server will authenticate the host by checking that the token signature matches the host public key that the client included in the token request)

BUG=115899

Review URL: https://chromiumcodereview.appspot.com/12313085