cc: Don't draw and swap if the frame will not change.
When there is no visible damage in the root surface, we have no
reason to do CalculateRenderPasses, DrawLayers, or SwapBuffers.

This adds an early out in each of these when there is no damage,
but storing a flag on the FrameData to communicate this state.

When doing a readback, we need to make sure we draw the area
being read back, so we pass a damage rect in to PrepareToDraw
to enforce this. This mechanism can be used to also implement
the "ForceFullFrameDamage" mechanism, so we move the flag to
LayerTreeHostImpl and have DamageTracker take a general rect
instead.

Before:
[ RUN      ] LayerTreeHostPerfTestJsonReader.TenTenSingleThread
*RESULT 10_10_layer_tree: frames= 3089.37 runs/s

After:
[ RUN      ] LayerTreeHostPerfTestJsonReader.TenTenSingleThread
*RESULT 10_10_layer_tree: frames= 4679.13 runs/s

When there's no damage, a full single-threaded commit+composite speeds
up about 50%.

Tests:
DamageTrackerTest.DamageWhenAddedExternally
LayerTreeHostDamageTestNoDamageDoesNotSwap.RunSingleThread
LayerTreeHostDamageTestNoDamageDoesNotSwap.RunMultiThread
LayerTreeHostDamageTestNoDamageReadbackDoesDraw.RunSingleThread
LayerTreeHostDamageTestNoDamageReadbackDoesDraw.RunMultiThread
LayerTreeHostDamageTestForcedFullDamage.RunSingleThread
LayerTreeHostDamageTestForcedFullDamage.RunMultiThread

Adding a new perf test with damage:
LayerTreeHostPerfTestJsonReader.TenTenSingleThread_FullDamageEachFrame
*RESULT 10_10_layer_tree: frames= 3233.98 runs/s


R=nduca
BUG=222915

Review URL: https://chromiumcodereview.appspot.com/12662021