Mac: Fix BMB commands triggering a standalone window instead of a sheet.
It was caused by -menu getting called before the view got added to the
window, which was creating the C++ controller with a nil NSWindow.
(This was being done by the localizer inspecting the view.)

Add an early return in that case - so the menu only gets created once
the view is attached to the window.

BUG=227388
TEST=As per bug.

Review URL: https://chromiumcodereview.appspot.com/13801018