Turn on component updater on chromeos, only for the pnacl component.
Make the pnacl component place files in a profile-specific directory
for chromeos.  We don't want to share data between users on
chromeos because of questions about integrity of these files.
This adds a notification watcher and will re-register pnacl with
a different version whenever the user logs into
a different account.

Files are installed to a stateful partition, which should
be okay.  The stateful partition is mounted no-exec, but
the pnacl files are not typical executables.

This is a temporary solution to help with testing out PNaCl on chromeos.
We will need to find a different solution later that does
not involve multiple copies, one for each user.

BUG=221381
TEST= manual with:

1) export GYP_DEFINES="chromeos=1"
2) build
3) out/Release/chrome --enable-pnacl --enable-nacl --user-data-dir=/tmp/temp_profile_chromeos --login-profile=user --login-manager
4) wait 20 seconds and see /tmp/temp_profile_chromeos/test-user/pnacl/0.0.0.5 show up

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=192415

Review URL: https://codereview.chromium.org/13071002