cc: Turn on presubmit cpplint warnings
cc is nearly spotless with cpplint warnings, so let's keep it that way.
The CommonChecks lint warning turns off a ton of errors that I would
prefer to keep on.  I turned it to level 1, since that's the default
when running cpplint.py.

This doesn't appear to be on by default.

R=jamesr@chromium.org
BUG=227196


Review URL: https://chromiumcodereview.appspot.com/13483022