History: Ignore server side de-duping of synced history; do it client side.
Results returned from the history server are de-duped so that a given URL
only appears once per day. However, this de-duping should be done based on
the client's timezone, which the server does not know. So create a separate
HistoryEntry for each visit, and do the de-duping on the client.

BUG=228875


Review URL: https://chromiumcodereview.appspot.com/13741009