drive: Extends fileBrowserPrivate.searchDriveMetadata
Adds |options| and |maxNumResults| parameters to the API,
and makes it return everything for an empty query.
A caller can set |option| to "FILES_ONLY" to have only files returned in the results.
The caller can also specify the maximum number of results as |maxNumResults|.

BUG=168741
TEST=unittests

Review URL: https://chromiumcodereview.appspot.com/12386013