Telemetry: Allow test to specify skipping the first iteration
For Startup tests, some test cases exercise "Warm" runs of Chrome.  To facilitate this, it's desirable to be able to perform an initial test run whose results are ignored and only serves to "warm up" resources needed by the test.

This CL adds a |discard_first_result| property to the |PageTest| object which serves this purpose.
BUG=


Review URL: https://chromiumcodereview.appspot.com/13773006