gpu: Disable virtual contexts for Qualcomm on SDK-16
We needed virtual contexts for Qualcomm, but we now
have a startup crasher after enabling composite-to-mailbox.
This disables virtual context on SDK-16 for beta until
we can find a better solution.

BUG=187487
NOTRY=true

No-try as this is android only and we need it for beta.

Review URL: https://chromiumcodereview.appspot.com/13913021