Create and show the virtual keyboard.
Introduces ui/keyboard/ which contains generic code for the control of the keyboard (right now, just creation of the container window).

Also includes the necessary changes to ash so that the keyboard is (always) displayed on ChromeOS builds when the --enable-virtual-keyboard flag is present.

BUG=226986

Review URL: https://codereview.chromium.org/13164002