CrOS: Look for consistency between successive vsync intervals.
Currently we report computed vsync intervals based on if they are in
an expected range and silently drop otherwise. The disadvantage with this
is that we're probably reporting noise caused by configuration changes
(monitor reconfiguration or moving a window between monitors). We may
also be papering over driver bugs.

Instead, we look for consistency between successive computed intervals
as a way of validating that there hasn't been a configuration change. In
this case, we can be more aggressive in error reporting to find driver bugs
(crash instead of silently drop).

BUG=221701


Review URL: https://chromiumcodereview.appspot.com/13861016