Prioritize tiles near the viewport in out-of-viewport layers
Not prioritizing tiles near the viewport in out-of-viewport layers causes
frequent invalidation and redraw of the tiles when the page is scrolled.

Modified ExpandRectEquallyToAreaBoundedBy() to expand starting_rect first
and check if further expansion is needed for the intersection of
starting_rect and bounding_rect.

BUG=223697

Review URL: https://chromiumcodereview.appspot.com/13582006